+++ CHANGELOG.rst
@@ -6,6 +6,10 @@ All notable changes to the kytos project will be documented in this file.
 UNRELEASED - Under development
 ******************************
 
+Changed
+=======
+- Changed ``UNI.is_valid`` to allow tags such as ``any`` and ``untagged``
+
 [2022.3.0]  2022-12-15
 **********************